fix(core): lift the Yee compute dtype off its float32 pin so field storage dtype actually reaches the update arithmetic (closes #630)
rfx/core/yee.py's update_h/update_e computed a compute dtype that was
hard-pinned to float32 for any real field storage dtype (`_fdtype =
state.ex.dtype` was used only to cast the *result* back). A float64
field carry was silently re-quantized to float32 at the top of every
timestep -- invisible to a leaf-level dtype census since the pinning
quantity was an in-step cast, not a closed-over array. The public API
had no way to reach float64 storage anyway, so this surfaced only once
FD-vs-AD gates on small, localized design perturbations still capped at
a ~1e-7 relative noise floor with sign-flipping, non-monotone scatter.

Fix: _cdtype is now jnp.promote_types(state.ex.dtype, jnp.float32) at
both update_h/update_e call sites, plus all 51 literal
`.astype(jnp.float32)` field-arithmetic sites across the NU, GPU-fast,
and anisotropic lanes. This is byte-identical on the production
float32 path (promote_types(float32, float32) == float32) and
preserves the existing float16 upcast intent (mixed precision)
unchanged -- verified via a falsifier: reverting only the two
_cdtype sites while keeping all 51 literal changes reproduces the
stock float64-fields result exactly.

Also adds Simulation(precision="float64") -- previously only
"float32"/"mixed" were valid, and forward() (the lane every FD-vs-AD
gate uses) never threaded field_dtype from self._precision at all, so
even "mixed" was unreachable from forward(). A new preflight check
warns if precision="float64" is requested without JAX x64 enabled,
since JAX otherwise silently downcasts back to float32 with no other
signal.

Corrects eight test docstrings/comments that attributed FD-vs-AD noise
floors to "float32 fields" (storage) when the actual mechanism was the
storage-independent Yee compute dtype. No existing gate/tolerance is
tightened in this PR.

R3: memory=docs/agent-memory/rfx-known-issues.md "#630 RESOLVED" entry (primary checkout, added this session) + project_issue527_f32_comparator + feedback_gate_can_bind_artifact + feedback_f64_referee_works_despite_f32_pin (all consistent -- none contradicted) | R2-attempts=0 (root cause and fix were fully diagnosed in the issue before this session; this is single-attempt execution of a pre-declared fix, not a repeated mechanism-equivalent retry) | falsifier=revert only the two _cdtype sites while keeping the 51 literal-site changes; confirmed the result returns to stock's float64-fields numbers on every checked quantity (g_ad, 4 FD-ladder rel_err%, 8 sub-signal trace deltas)

+++ tests/test_mixed_precision.py
@@ -88,7 +88,7 @@ class TestSimulationMixedPrecision:
     def test_precision_parameter_validation(self):
         with pytest.raises(ValueError, match="precision"):
             Simulation(freq_max=5e9, domain=(0.02, 0.02, 0.02),
-                       boundary="pec", precision="float64")
+                       boundary="pec", precision="float128")
 
     def test_mixed_precision_runs(self):
         """Mixed precision simulation completes without NaN."""
